Rename viz/witness.py to viz/validator.py with Validator/Validators classes
